Rename java package to align with product name
I bumped the maven snapshot version to 4.0.0.
