+++ src/main/java/com/flashcards4u/controller/DeckController.java
@@ -1,9 +1,9 @@
-package com.chrisbrickey.flashcards3.controller;
+package com.flashcards4u.controller;
 
-import com.chrisbrickey.flashcards3.model.Card;
-import com.chrisbrickey.flashcards3.response.CardResponse;
-import com.chrisbrickey.flashcards3.response.DeckResponse;
-import com.chrisbrickey.flashcards3.service.DeckService;
+import com.flashcards4u.model.Card;
+import com.flashcards4u.response.CardResponse;
+import com.flashcards4u.response.DeckResponse;
+import com.flashcards4u.service.DeckService;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
