+++ src/test/java/com/flashcards4u/Flashcards4uApplicationTests.java
@@ -1,10 +1,10 @@
-package com.chrisbrickey.flashcards3;
+package com.flashcards4u;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class Flashcards3ApplicationTests {
+class Flashcards4uApplicationTests {
 
 	@Test
 	void contextLoads() {
