+++ src/test/java/com/flashcards4u/SmokeTest.java
@@ -1,7 +1,7 @@
-package com.chrisbrickey.flashcards3;
+package com.flashcards4u;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import com.chrisbrickey.flashcards3.controller.RootController;
+import com.flashcards4u.controller.RootController;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
