+++ src/test/java/com/flashcards4u/response/CardResponseTest.java
@@ -1,4 +1,4 @@
-package com.chrisbrickey.flashcards3.response;
+package com.flashcards4u.response;
 
 import org.junit.jupiter.api.Test;
