+++ src/test/java/com/flashcards4u/service/DeckServiceTest.java
@@ -1,6 +1,6 @@
-package com.chrisbrickey.flashcards3.service;
+package com.flashcards4u.service;
 
-import com.chrisbrickey.flashcards3.model.Card;
+import com.flashcards4u.model.Card;
 import org.junit.jupiter.api.Test;
 
 import java.util.Collections;
